Move `polar_plot.py` to root of `plot` dir
