+++ e3sm_diags/driver/polar_driver.py
@@ -15,7 +15,7 @@ from e3sm_diags.driver.utils.regrid import (
 from e3sm_diags.driver.utils.type_annotations import MetricsDict
 from e3sm_diags.logger import custom_logger
 from e3sm_diags.metrics.metrics import correlation, rmse, spatial_avg
-from e3sm_diags.plot.cartopy.polar_plot import plot as plot_func
+from e3sm_diags.plot.polar_plot import plot as plot_func
 
 logger = custom_logger(__name__)
